backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
asgiref
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
bitarray==2.9.2
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
channels
channels_redis
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.5
cytoolz==0.12.3
daphne
django
djangorestframework
django-celery-beat==2.6.0
django-cors-headers
django-extensions
django-timezone-field==6.1.0
eth-account==0.11.0
eth-hash==0.7.0
eth-keyfile==0.8.0
eth-keys==0.5.0
eth-rlp
eth-typing==4.0.0
eth-utils==4.0.0
eth_abi==5.0.0
frozenlist==1.4.1
hexbytes==0.3.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu
lru-dict==1.2.0
multidict==6.0.5
parsimonious==0.9.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psycopg2-binary
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pyOpenSSL==24.0.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2024.1
pyunormalize==15.1.0
redis==5.0.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rlp==4.0.0
rpds-py==0.18.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
sqlalchemy==2.0.29
toolz==0.12.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions
tzdata==2024.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
web3==6.15.1
websockets==12.0
yarl==1.9.4
zope.interface==6.2



